requirements.txt
discord.py>=2.3.0,<3.0.0
aiohttp>=3.9.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
